* added extra return types and delete file function

+++ software/cobalt_ant_bringup/inc/logfs.h
@@ -6,7 +6,7 @@
 result fsInit();
 
 result fsFileCreate(uint16_t fileId, uint32_t fileSize);
-result fsFileDelete(uint32_t fileId);
+result fsFileDelete(uint16_t fileId);
 // TODO File listing
 // first file in dir
 // next file in dir
